Added support for jpg/bmp texture loading; OBJ loading enhanced with support of scenes. Like sponza; Added new option to import OBJ files with this option user can load multiple objects out of OBJ(meshes, materials...); LoadPNGHeightmap now checks whether png is 16 bit gray scale;Fixed heap corruption in FETerrain::updateCPUHeightInfo(); Add support for 64 bit version of project. Add some lib versions for 64 bits; Also added option to import glTF file format;Added filter by type in content browser; Unified all select sub windows to one that can work with all FEObjects; GPU_CULLING is now only one way of culling left for instanced entities;Fixed range calculations on gameModel edit window loading; update glfw to 3.3.7 version; update imgui to 1.8.7 version; moved glfw, glew2 and imgui abstraction to FEBasicApplication class. And add it as a submodule; Fixed tangent calculations in terrain shader and in OBJ loader; Fixed all tangent values in meshes data in all projects.

+++ Renderer/FELight.cpp
@@ -355,7 +355,7 @@ void FEDirectionalLight::updateCascades(float cameraFov, float aspectRatio, floa
 			float   t;
 
 			glm::mat4 cliping = cascadeData[i].projectionMat * cascadeView;
-			for (size_t j = 0; j < 4; j++)
+			for (int j = 0; j < 4; j++)
 			{
 				clip[j * 4] = cliping[j][0];
 				clip[j * 4 + 1] = cliping[j][1];
